+++ src/test/mc/StructurePoolBlockPredicateBlockMatchRandom.h
@@ -23,26 +23,21 @@ public:
     // NOLINTEND
 
 public:
-    StructurePoolBlockPredicateBlockMatchRandom(::Block const& block, float probability)
+    StructurePoolBlockPredicateBlockMatchRandom(::Block const& block, float const probability)
     : mBlock(block),
-      mProbability(probability){};
+      mProbability(probability) {}
 
 public:
     // virtual functions
     // NOLINTBEGIN
-    // vIndex: 2
     virtual bool test(::Block const& block, ::Randomize& randomize) const /*override*/;
 
-    // vIndex: 1
     virtual bool test(::BlockPos const& worldPos, ::BlockPos const& refPos, ::Randomize& randomize) const /*override*/;
 
-    // vIndex: 5
     virtual ::StructurePoolBlockPredicateType getType() const /*override*/;
 
-    // vIndex: 6
     virtual void appendMetadataKey(::Util::XXHash& hash) const /*override*/;
 
-    // vIndex: 0
     virtual ~StructurePoolBlockPredicateBlockMatchRandom() /*override*/ = default;
     // NOLINTEND