+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.0] - 2026-01-28
+
+### Changed
+
+- Supported LeviLamina 1.9.2(Client 1.21.132.1 and Server 1.21.132.3)
+- Resolved some warnings in code
+
 ## [0.13.0-rc.1] - 2026-01-20
 
 ### Added
@@ -162,7 +169,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [#21]: https://github.com/LiteLDev/MoreDimensions/issues/21
 
-[Unreleased]: https://github.com/LiteLDev/MoreDimensions/compare/v0.13.0-rc.1...HEAD
+[Unreleased]: https://github.com/LiteLDev/MoreDimensions/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/LiteLDev/MoreDimensions/compare/v0.13.0-rc.1...v0.13.0
 [0.13.0-rc.1]: https://github.com/LiteLDev/MoreDimensions/compare/v0.12.2...v0.13.0-rc.1
 [0.12.2]: https://github.com/LiteLDev/MoreDimensions/compare/v0.12.1...v0.12.2
 [0.12.1]: https://github.com/LiteLDev/MoreDimensions/compare/v0.12.0...v0.12.1